Sentiment Graph Builder
TSQL code to quickly build a binary graph by averages for each month (in
this case, the binary pieces are positive and negative - this could
changed to anything else that's binary).  This is based off a table that
is then pivoted by month, with the binary data inputted as the average.